requirements.txt
numpy==1.26.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
pyserial==3.5
